requirements.txt
pandas
pytorch
numpy
torchvision
nibabel
matplotlib
glob
wandb
logging
tqdm